Move edit input/output to MatChipRow.
